requirements.txt
torch
torchvision
bytecode
pillow
matplotlib
opencv-python
torchviz
gradio
bitsandbytes
accelerate
git+https://github.com/openai/CLIP.git
pycocotools
numpy
# Optional dependency for improved panoptic mask handling
panopticapi
